Committing the whole project at once
This project accepts 4 parameters - CSV file path, sender email, password for the respective email and recipient email. The CSV file holds an information about the weather forecast for a given time period. Then the application generates a new CSV file with an aggregates (such as min/max/average/median value) of the given data as well as the most appropriate dates for space launch (based on a specific criteria). Afterwards the application send an email from the "sender email", authenticating them with the given "password", attaches the newly created CSV file and sends the email to the given "recipient email".